Downloads/crypto_trading_bot/requirements.txt
python-binance
python-dotenv
openai
PyQt5
requests
plotly
dash
bs4
feedparser
